2nd change

+++ assets/js/index.js
@@ -3,6 +3,10 @@ $("#add_user").submit(function(event){
     alert("Data Inserted Successfully!");
 })
 
+$("#add_dossier").submit(function(event){
+    alert("Data Inserted Successfully!");
+})
+
 $("#update_user").submit(function(event){
     event.preventDefault();
 
@@ -33,6 +37,50 @@ if(window.location.pathname == "/"){
 
         var request = {
             "url" : `http://localhost:5000/api/users/${id}`,
+            "method" : "DELETE"
+        }
+
+        if(confirm("Do you really want to delete this record?")){
+            $.ajax(request).done(function(response){
+                alert("Data Deleted Successfully!");
+                location.reload();
+            })
+        }
+
+    })
+}
+
+/******************* */
+$("#update_dossier").submit(function(event){
+    event.preventDefault();
+
+    var unindexed_array = $(this).serializeArray();
+    var dat = {}
+
+    $.map(unindexed_array, function(n, i){
+        data[n['name']] = n['value']
+    })
+
+
+    var request = {
+        "url" : `http://localhost:5000/api/dossiers/${data.id}`,
+        "method" : "PUT",
+        "data" : data
+    }
+
+    $.ajax(request).done(function(response){
+        alert("Data Updated Successfully!");
+    })
+
+})
+
+if(window.location.pathname == "/dossiers"){
+    $ondelete = $(".table tbody td a.delete");
+    $ondelete.click(function(){
+        var id = $(this).attr("data-id")
+
+        var request = {
+            "url" : `http://localhost:5000/api/dossiers/${id}`,
             "method" : "DELETE"
         }
